refacotr: Apply WPS412 and others 1. Apply WPS412 rule 2. Change minimum python version from 3.9 to 3.10 3. Refactor dockerfile
