+++ app/data/prompt/__init__.py
@@ -1,7 +1,4 @@
-import sys
-from pathlib import Path
-
-from data.prompt.system_prompts import *
 from data.prompt.answer_prompts import *
 from data.prompt.question_prompts import *
 from data.prompt.search_prompts import *
+from data.prompt.system_prompts import *
