+++ app/modules/__init__.py
@@ -1,0 +1,3 @@
+from modules import chatbot, client, database, feedback
+
+__all__ = ["chatbot", "client", "database", "feedback"]
