+++ app/modules/chatbot/test/__init__.py
@@ -1,0 +1,1 @@
+from modules.chatbot.test.stream import *
