+++ app/modules/client/__init__.py
@@ -1,6 +1,2 @@
-import sys
-from pathlib import Path
-
-(str(Path(__file__).parents[3]))
 from modules.client.openai_client import *
 from modules.client.thanosql_client import *
