+++ app/router/__init__.py
@@ -1,4 +1,2 @@
-from router.settings import settings_router
 from router.defaults import default_router
-
-__all__ = ["default_router", "settings_router"]
+from router.settings import settings_router
