+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = ["KyungHo Kim <hao@smartmind.team>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.14"
+python = ">=3.10,<3.14"
 langchain-openai = "^0.2.12"
 langchain-postgres = "^0.0.12"
 langchain-core = "^0.3.25"
@@ -23,7 +23,6 @@ redis = "^5.2.1"
 tabulate = "^0.9.0"
 thanosql = "^1.0.11"
 psycopg2-binary = "^2.9.10"
-
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.4"
